- integrated channel handler which can run JBehave scripts

+++ lib/netty-bgp4/src/main/java/org/bgp4j/netty/protocol/open/OpenPacket.java
@@ -89,7 +89,7 @@ public class OpenPacket extends BGPv4Packet {
 	/**
 	 * @param bgpIdentifier the bgpIdentifier to set
 	 */
-	public void setBgpIdentifier(int bgpIdentifier) {
+	public void setBgpIdentifier(long bgpIdentifier) {
 		this.bgpIdentifier = bgpIdentifier;
 	}
